(V-Hw01.00.00-Fw01.02.00)新增网络控制功能：外部程序经以太网LCM发送rc_control_command_lcmt消息（通道rc_control_command，字段mode/gait_type/v_des/omega_des）等效手柄按键控制机器人；手柄任意按键按下即夺回控制权；手柄原有控制逻辑不变。    注意事项：a) 网络端按离散指令发送（每条消息均会重新接管控制权，连续流式发送会阻止手柄夺回）；b) 夺回会同时执行该按键自身动作：按A夺回会顺带切换站立/行走，按B/RB会进入校准模式——LOCOMOTION中无副作用的夺回键为单独按LB或START；c) 手柄缺失或中途拔出时无法夺回，且无网络活性超时，行走中需保证网络端或手柄至少一方可用；d) 板上applied日志仅在模式变化时打印，同模式内更新gait/速度静默接受（以机器人行为或lcm-spy确认）。

+++ EngineAI_Controller/lcm-types/lcmtypes/control_parameter_respones_lcmt.java
@@ -85,11 +85,11 @@ public final class control_parameter_respones_lcmt implements lcm.lcm.LCMEncodab
     public void _decodeRecursive(DataInput ins) throws IOException
     {
         this.name = new byte[(int) 64];
-        ins.readFully(this.name, 0, 64); 
+        ins.readFully(this.name, 0, (int) 64); 
         this.requestNumber = ins.readLong();
  
         this.value = new byte[(int) 64];
-        ins.readFully(this.value, 0, 64); 
+        ins.readFully(this.value, 0, (int) 64); 
         this.parameterKind = ins.readByte();
  
         this.requestKind = ins.readByte();
